Lab5/cdk-pycon-flask/requirements.txt
-e .
attrs==19.1.0
autopep8==1.4.4
aws-cdk.assets==1.8.0
aws-cdk.aws-apigateway==1.8.0
aws-cdk.aws-applicationautoscaling==1.8.0
aws-cdk.aws-autoscaling==1.8.0
aws-cdk.aws-autoscaling-common==1.8.0
aws-cdk.aws-autoscaling-hooktargets==1.8.0
aws-cdk.aws-certificatemanager==1.8.0
aws-cdk.aws-cloudformation==1.8.0
aws-cdk.aws-cloudfront==1.8.0
aws-cdk.aws-cloudwatch==1.8.0
aws-cdk.aws-codebuild==1.8.0
aws-cdk.aws-codecommit==1.8.0
aws-cdk.aws-codepipeline==1.8.0
aws-cdk.aws-ec2==1.8.0
aws-cdk.aws-ecr==1.8.0
aws-cdk.aws-ecr-assets==1.8.0
aws-cdk.aws-ecs==1.8.0
aws-cdk.aws-ecs-patterns==1.8.0
aws-cdk.aws-eks==1.8.0
aws-cdk.aws-elasticloadbalancing==1.8.0
aws-cdk.aws-elasticloadbalancingv2==1.8.0
aws-cdk.aws-events==1.8.0
aws-cdk.aws-events-targets==1.8.0
aws-cdk.aws-iam==1.8.0
aws-cdk.aws-kms==1.8.0
aws-cdk.aws-lambda==1.8.0
aws-cdk.aws-logs==1.8.0
aws-cdk.aws-route53==1.8.0
aws-cdk.aws-route53-targets==1.8.0
aws-cdk.aws-s3==1.8.0
aws-cdk.aws-s3-assets==1.8.0
aws-cdk.aws-secretsmanager==1.8.0
aws-cdk.aws-servicediscovery==1.8.0
aws-cdk.aws-sns==1.8.0
aws-cdk.aws-sns-subscriptions==1.8.0
aws-cdk.aws-sqs==1.8.0
aws-cdk.aws-ssm==1.8.0
aws-cdk.aws-stepfunctions==1.8.0
aws-cdk.core==1.8.0
aws-cdk.custom-resources==1.8.0
aws-cdk.cx-api==1.8.0
aws-cdk.region-info==1.8.0
cattrs==0.9.0
jsii==0.16.0
publication==0.0.3
pycodestyle==2.5.0
python-dateutil==2.8.0
six==1.12.0
typing-extensions==3.7.4
